requirements.txt
feedparser
pyyaml